Added 2 algorithms
the search with instructions one and the one that will put all the
receipes in alphabetical order tho... as long as the recipes do not have
ingredients attached to them the first algorithm cannot work as I need
to access all the ingredients from that recipe to be able to compare it
to what the user wants. unless you have a link from an ingredient to the
recipes it is in but that would make the algorithm a whole lot more
complex than it already is.